+++ Assets/OxGFrame/CoreFrame/Example/UIFrameDemo/Scripts/Demo1UI.cs
@@ -61,12 +61,12 @@ public class Demo1UI : UIBase
         */
     }
 
-    protected override void ShowAnim(AnimEndCb animEndCb)
+    protected override void ShowAnime(AnimeEndCb animEndCb)
     {
         animEndCb(); // Must Keep, Because Parent Already Set AnimCallback
     }
 
-    protected override void HideAnim(AnimEndCb animEndCb)
+    protected override void HideAnime(AnimeEndCb animEndCb)
     {
         animEndCb(); // Must Keep, Because Parent Already Set AnimCallback
     }