+++ Assets/OxGFrame/CoreFrame/Example/UIFrameDemo/Scripts/Demo3UI.cs
@@ -60,12 +60,12 @@ public class Demo3UI : UIBase
         */
     }
 
-    protected override void ShowAnim(AnimEndCb animEndCb)
+    protected override void ShowAnime(AnimeEndCb animEndCb)
     {
         animEndCb(); // Must Keep, Because Parent Already Set AnimCallback
     }
 
-    protected override void HideAnim(AnimEndCb animEndCb)
+    protected override void HideAnime(AnimeEndCb animEndCb)
     {
         animEndCb(); // Must Keep, Because Parent Already Set AnimCallback
     }